requirements.txt
anthropic>=0.40.0
requests>=2.31.0
feedparser>=6.0.11
pytrends>=4.9.2
python-dotenv>=1.0.0
